// src/app/models/photographer/photographer.mode.ts
export interface PhotographerProps {
  id?: number
  name: string
  email: string
  cpf: string
  gender: string
  phone: string
  zipCode: string
  state: string
  city: string
  address: string
  neighborhood: string
  number: string
  complement: string
  password: string
  confirmPassword: string
}